refactor: move claudecli to internal/providers/claudecli
