+++ experimental/adk-go/internal/providers/claudecli/adapter.go
