+++ experimental/adk-go/internal/providers/claudecli/adapter_test.go
